+++ src/ui/main_window.py
@@ -11,6 +11,7 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
+from data.database import application_path
 from ui.input_forms import InputForms
 
 logger = logging.getLogger("ui")
@@ -21,9 +22,14 @@ class MainWindow(QMainWindow):
         self.internet_connection_status = internet_connection_status
         super().__init__(parent)
 
-        self.setWindowTitle("Agricultural Carbon Calculator")
+        self.setWindowTitle("Carbon Calculator")
         self.setGeometry(100, 100, 400, 300)
-        self.setWindowIcon(QIcon("resources/assets/icon.png"))
+        if sys.platform == "darwin":
+            self.setWindowIcon(QIcon(f"{application_path}/resources/assets/icon.icns"))
+        elif sys.platform == "win32":
+            self.setWindowIcon(QIcon(f"{application_path}/resources/assets/icon.ico"))
+        else:
+            self.setWindowIcon(QIcon(f"{application_path}/resources/assets/icon.png"))
         # TODO: find a way to change the icon on the taskbar
 
         central_widget = QWidget()
@@ -50,27 +56,23 @@ class MainWindow(QMainWindow):
         else:
             event.ignore()
 
-    def display_internet_connection_status(self, internet_connection_status):
-        logger.info(f"Internet connection status: {internet_connection_status}")
-        internet_connection_status_msg_box = QMessageBox(self)
-        internet_connection_status_msg_box.setWindowTitle(
-            f"Internet Connection Status: {internet_connection_status}"
-        )
+    def display_internet_connection_status(self, has_internet):
+        """Display a message box with the internet connection status."""
+        msg_box = QMessageBox(self)
+        msg_box.setWindowTitle("Status")
         ok_button = QPushButton("Ok.")
-        internet_connection_status_msg_box.addButton(ok_button, QMessageBox.AcceptRole)
-        if internet_connection_status is False:
-            internet_connection_status_msg_box.setText(
+        msg_box.addButton(ok_button, QMessageBox.AcceptRole)
+
+        if has_internet:
+            msg_box.setText("Internet connection found.")
+        elif has_internet is False:
+            msg_box.setText(
                 "No internet connection found. Some features may not be available."
             )
-            internet_connection_status_msg_box.exec_()
-        elif internet_connection_status is True:
-            internet_connection_status_msg_box.setText("internet connection found.")
-            internet_connection_status_msg_box.exec_()
         else:
-            internet_connection_status_msg_box.setText(
-                "There was an error checking the internet connection."
-            )
-            internet_connection_status_msg_box.exec_()
+            msg_box.setText("There was an error checking the internet connection.")
+
+        msg_box.exec_()
 
 
 if __name__ == "__main__":